rename truncated exp merge name into constant
