+++ vlgp/constant.py
@@ -1,18 +1,43 @@
 """constants"""
 
-# coding observation dimensions
 from collections import defaultdict
 
-from .name import Z_DIM, Y
+# names
+VB = 'VB'
+MAP = 'MAP'
+X = 'x'
+Z = 'z'
+MU = 'mu'
+W = 'w'
+V = 'v'
+dMU = 'dmu'
+dA = 'da'
+dB = 'db'
+ESTEP = 'learn_post'
+MSTEP = 'learn_param'
+HSTEP = 'learn_hyper'
+HPERIOD = 'nhyper'
+HOBJ = 'hyper_obj'
+TRIALLET = 'subsample_size'
+PRIOR = 'chol'
+PRIORSTD = 'sigma'
+PRIORTIMESCALE = 'omega'
+MAXITER = 'niter'
+ITER = 'it'
+LIK = 'lik'
+Y_DIM = 'y_dim'
+#
 
 NA = 0  # dimension not to use
-POISSON = 1  # spike train
-GAUSSIAN = 2  # local field potential
+POISSON = 1
+GAUSSIAN = 2
 
 LIK_CODE = defaultdict(int,
                        poisson=POISSON,
                        gaussian=GAUSSIAN)
 
+Y = 'y'
+Z_DIM = 'lat_dim'
 REQUIRED_FIELDS = [Y, Z_DIM]
 
 # default options
@@ -47,4 +72,5 @@ DEFAULT_VALUES = {'learning_rate': 1.0,
                   # misc
                   'verbose': False,
                   'saving_interval': 3600}
+
