--- vlgp/name.py
@@ -1,33 +1,0 @@
-"""keys"""
-
-VB = 'VB'
-MAP = 'MAP'
-
-# field names
-Y = 'y'
-X = 'x'
-Z = 'z'
-MU = 'mu'
-W = 'w'
-V = 'v'
-dMU = 'dmu'
-dA = 'da'
-dB = 'db'
-
-ESTEP = 'learn_post'
-MSTEP = 'learn_param'
-HSTEP = 'learn_hyper'
-HPERIOD = 'nhyper'
-HOBJ = 'hyper_obj'
-TRIALLET = 'subsample_size'
-PRIOR = 'chol'
-PRIORSTD = 'sigma'
-PRIORTIMESCALE = 'omega'
-
-MAXITER = 'niter'
-ITER = 'it'
-
-LIK = 'lik'
-
-Z_DIM = 'lat_dim'
-Y_DIM = 'y_dim'
